Add support for organization header
https://beta.openai.com/docs/api-reference/requesting-organization

Copilot generated the WithXXX, the conditional in makeHTTPRequest method and
interestingly knew the correct header name.
