+++ openai.go
@@ -1,18 +1,10 @@
 package openai
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"io"
-	"math/rand"
-	"mime/multipart"
 	"net/http"
-	"os"
-	"path/filepath"
 	"runtime"
-	"strconv"
 )
 
 const (
@@ -50,6 +42,9 @@ type openAI struct {
 	APIKey string
 	// Client is the http client to use to make requests to the OpenAI API
 	Client HttpClient
+	// organization is the organization to use for the requests to the OpenAI API.
+	// See https://beta.openai.com/docs/api-reference/requesting-organization
+	organization string
 }
 
 type openAIOption func(*openAI)
@@ -59,6 +54,14 @@ type openAIOption func(*openAI)
 func WithHttpClient(client HttpClient) openAIOption {
 	return func(o *openAI) {
 		o.Client = client
+	}
+}
+
+// WithOrganization sets the organization to use for the requests to the OpenAI
+// API. See https://beta.openai.com/docs/api-reference/requesting-organization
+func WithOrganization(organization string) openAIOption {
+	return func(o *openAI) {
+		o.organization = organization
 	}
 }
 
@@ -72,82 +75,4 @@ func NewOpenAI(apiKey string, options ...openAIOption) OpenAI {
 		option(res)
 	}
 	return res
-}
-
-func (o *openAI) makeJSONRequest(ctx context.Context, uri string, req any, resp any) error {
-	bodyBytes, err := json.Marshal(req)
-	if err != nil {
-		return fmt.Errorf("openai: JSON encoding error: %w", err)
-	}
-	body := bytes.NewBuffer(bodyBytes)
-	httpReq, err := http.NewRequestWithContext(ctx, "POST", uri, body)
-	if err != nil {
-		return fmt.Errorf("openai: HTTP request creation error: %w", err)
-	}
-
-	httpReq.Header.Set("Content-Type", "application/json")
-	return o.makeHttpRequest(httpReq, resp)
-}
-
-// makeMultiPartRequest makes a multipart request to the OpenAI API. It accepts a
-// map of form fields and a list of files paths to upload.
-func (o *openAI) makeMultiPartRequest(ctx context.Context, uri string, fields map[string]string, files map[string]string, resp any) error {
-	body := &bytes.Buffer{}
-	writer := multipart.NewWriter(body)
-	for key, val := range fields {
-		if err := writer.WriteField(key, val); err != nil {
-			return fmt.Errorf("openai: multipart form field encoding error: %w", err)
-		}
-	}
-	for key, file := range files {
-		fileWriter, err := writer.CreateFormFile(key, filepath.Base(file))
-		if err != nil {
-			return fmt.Errorf("openai: multipart form file encoding error: %w", err)
-		}
-		fh, err := os.Open(file)
-		if err != nil {
-			return fmt.Errorf("openai: multipart form file opening error: %w", err)
-		}
-		defer fh.Close()
-		if _, err = io.Copy(fileWriter, fh); err != nil {
-			return fmt.Errorf("openai: multipart form file copying error: %w", err)
-		}
-	}
-	if err := writer.Close(); err != nil {
-		return fmt.Errorf("openai: multipart form closing error: %w", err)
-	}
-	httpReq, err := http.NewRequestWithContext(ctx, "POST", uri, body)
-	if err != nil {
-		return fmt.Errorf("openai: HTTP request creation error: %w", err)
-	}
-	httpReq.Header.Set("Content-Type", writer.FormDataContentType())
-	return o.makeHttpRequest(httpReq, resp)
-}
-
-func (o *openAI) makeHttpRequest(httpReq *http.Request, resp any) error {
-	httpReq.Header.Set("Authorization", "Bearer "+o.APIKey)
-	httpReq.Header.Set("User-Agent", userAgent)
-	httpReq.Header.Set("Accept", "application/json")
-	httpReq.Header.Set("X-Request-ID", "openai-go-"+strconv.FormatUint(rand.Uint64(), 16))
-	httpResp, err := o.Client.Do(httpReq)
-	if err != nil {
-		return fmt.Errorf("openai: HTTP error: %w", err)
-	}
-
-	if err = checkErrResponse(httpResp); err != nil {
-		return err
-	}
-	defer httpResp.Body.Close()
-
-	if err = checkJSONContentType(httpResp); err != nil {
-		return err
-	}
-	responseBody, err := io.ReadAll(httpResp.Body)
-	if err != nil {
-		return fmt.Errorf("openai: HTTP response read error: %w", err)
-	}
-	if err := json.Unmarshal(responseBody, resp); err != nil {
-		return fmt.Errorf("openai: HTTP success response JSON decoding error: %w", err)
-	}
-	return nil
 }
